Extract membership info as helper function, use in PayPalSuccess
