+++ src/components/PayPalSuccess.js
@@ -3,12 +3,14 @@ import { connect } from 'react-redux'
 import executeBillingAgreement from '../actions/executeBillingAgreement'
 import { Header, Container, Segment } from 'semantic-ui-react'
 import queryString from 'query-string'
+import membership from '../helpers/membershipInfo'
 import ErrorBoundary from './ErrorBoundary'
 import '../assets/PayPalSuccess.css'
 
 export const PayPalSuccess = props => {
   const params = queryString.parse(props.location.search)
   const [error, setError] = useState(false)
+  let name = membership(props, queryString).name
   useEffect(() => {
     if (props.error.length) {
       setError(true)
@@ -23,12 +25,11 @@ export const PayPalSuccess = props => {
           ? <Segment padded='very' className='payment-complete' raised>
             <Header as='h2' textAlign='center'>
               Thanks, you're now an AgileVentures {}
-              {params.plan.charAt(0).toUpperCase() + params.plan.slice(1)}{' '}
+              {name}{' '}
               Member!
             </Header>
             <Header as='h4' textAlign='center'>
-              Your 7 day free trial has now started. Your card will not be
-              charged until 7 days have passed.
+              {name === 'Premium' ? 'Your 7 day free trial has now started. Your card will not be charged until 7 days have passed.' : null}
             </Header>
             <Header as='h5' textAlign='center'>
               An AgileVentures mentor will be in touch shortly to help you
